试课管理联调+课程表下拉框

+++ hotdog-saas-tob/hotdog-saas-application/src/main/java/com/hotdog/saas/application/processor/education/clazz/trail/EducationCourseClassTrailDeleteProcessor.java
@@ -3,6 +3,10 @@ package com.hotdog.saas.application.processor.education.clazz.trail;
 import com.hotdog.saas.application.entity.request.education.clazz.trail.DeleteEducationCourseClassTrailRequest;
 import com.hotdog.saas.application.entity.response.BaseResponse;
 import com.hotdog.saas.domain.enums.ResultCodeEnum;
+import com.hotdog.saas.domain.exception.BusinessException;
+import com.hotdog.saas.domain.model.EducationCourseClassEnroll;
+import com.hotdog.saas.domain.model.EducationCourseClassTrail;
+
 import org.springframework.stereotype.Component;
 
 import lombok.extern.slf4j.Slf4j;
@@ -23,6 +27,13 @@ public class EducationCourseClassTrailDeleteProcessor extends AbstractEducationC
     public void doExecute(DeleteEducationCourseClassTrailRequest request, BaseResponse<Boolean> response) {
         Long id = request.getId();
         super.existsByClassTrailId(id);
+
+        EducationCourseClassTrail educationCourseClassTrail = educationCourseClassTrailRepository.findById(id);
+
+        if(educationCourseClassTrail.isAssign()){
+            throw new BusinessException("当前试课记录已分配班级，不能删除");
+        }
+
         Integer modifyFlag = educationCourseClassTrailRepository.remove(id, request.getOperator());
         response.setData(checkFlag(modifyFlag));
     }
